+++ lib/body_part_selector.dart
@@ -6,3 +6,5 @@ export 'src/body_part_selector_turnable.dart';
 export 'src/model/body_part_marker.dart';
 export 'src/model/body_parts.dart';
 export 'src/model/body_side.dart';
+export 'src/model/body_part_id.dart';
+export 'src/model/body_parts_extension.dart';